move migration code out of shared billing files
