+++ packages/zambdas/test/unit/billing/create-billing-claim-from-encounter.test.ts
@@ -66,10 +66,8 @@ import {
 } from '../../../src/billing/create-billing-claim-from-encounter/handler';
 import { validateRequestParameters } from '../../../src/billing/create-billing-claim-from-encounter/validateRequestParameters';
 import {
-  addClinicalPatientIdentifiers,
   BILLING_WORKING_COPY_TAG,
   buildNoCoverageStub,
-  clinicalFriendlyIdIdentifier,
   clinicalPatientIdentifier,
   CURRENT_STATUS_TAG_SYSTEM,
   EXCLUDE_WORKING_COPIES_PARAMS,
@@ -3859,173 +3857,6 @@ describe('create-billing-claim-from-encounter', () => {
       } as unknown as Oystehr;
 
       expect((await findMainBillingPatient(oystehr, 'patient-123')).total).toEqual(2);
-    });
-  });
-
-  describe('addClinicalPatientIdentifiers', () => {
-    it('adds the friendly id alongside the clinical patient id in one patch', async () => {
-      const patch = vi.fn().mockResolvedValue({});
-      await addClinicalPatientIdentifiers({
-        oystehr: {
-          fhir: {
-            patch,
-          },
-        } as unknown as Oystehr,
-        patient: billingResources.patient,
-        clinicalId: 'patient-123',
-        clinicalFriendlyId: '1015',
-      });
-
-      expect(patch).toHaveBeenCalledWith(
-        expect.objectContaining({
-          operations: [
-            {
-              op: 'add',
-              path: '/identifier',
-              value: [clinicalPatientIdentifier('patient-123'), clinicalFriendlyIdIdentifier('1015')],
-            },
-          ],
-        }),
-        expect.anything()
-      );
-    });
-
-    it('adds only the identifier that is missing', async () => {
-      const patch = vi.fn().mockResolvedValue({});
-      await addClinicalPatientIdentifiers({
-        oystehr: {
-          fhir: {
-            patch,
-          },
-        } as unknown as Oystehr,
-        patient: {
-          ...billingResources.patient,
-          identifier: [clinicalPatientIdentifier('patient-123')],
-        },
-        clinicalId: 'patient-123',
-        clinicalFriendlyId: '1015',
-      });
-
-      expect(patch).toHaveBeenCalledWith(
-        expect.objectContaining({
-          operations: [
-            {
-              op: 'add',
-              path: '/identifier/-',
-              value: clinicalFriendlyIdIdentifier('1015'),
-            },
-          ],
-        }),
-        expect.anything()
-      );
-    });
-
-    it('adds the identifier to a patient that has none', async () => {
-      const patch = vi.fn().mockResolvedValue({});
-      await addClinicalPatientIdentifiers({
-        oystehr: {
-          fhir: {
-            patch,
-          },
-        } as unknown as Oystehr,
-        patient: billingResources.patient,
-        clinicalId: 'patient-123',
-      });
-
-      expect(patch).toHaveBeenCalledWith(
-        {
-          resourceType: 'Patient',
-          id: 'billing-patient-123',
-          operations: [
-            {
-              op: 'add',
-              path: '/identifier',
-              value: [clinicalPatientIdentifier('patient-123')],
-            },
-          ],
-        },
-        {
-          optimisticLockingVersionId: '7',
-        }
-      );
-    });
-
-    it('appends rather than replacing when the patient carries other identifiers', async () => {
-      const patch = vi.fn().mockResolvedValue({});
-      await addClinicalPatientIdentifiers({
-        oystehr: {
-          fhir: {
-            patch,
-          },
-        } as unknown as Oystehr,
-        patient: {
-          ...billingResources.patient,
-          identifier: [
-            {
-              system: 'https://fhir.ottehr.com/other',
-              value: 'other',
-            },
-          ],
-        },
-        clinicalId: 'patient-123',
-      });
-
-      expect(patch).toHaveBeenCalledWith(
-        expect.objectContaining({
-          operations: [
-            {
-              op: 'add',
-              path: '/identifier/-',
-              value: clinicalPatientIdentifier('patient-123'),
-            },
-          ],
-        }),
-        {
-          optimisticLockingVersionId: '7',
-        }
-      );
-    });
-
-    it('writes nothing when the identifier is already there', async () => {
-      const patch = vi.fn();
-      await addClinicalPatientIdentifiers({
-        oystehr: {
-          fhir: {
-            patch,
-          },
-        } as unknown as Oystehr,
-        patient: {
-          ...billingResources.patient,
-          identifier: [clinicalPatientIdentifier('patient-123')],
-        },
-        clinicalId: 'patient-123',
-      });
-
-      expect(patch).not.toHaveBeenCalled();
-    });
-
-    it('does not re-add the identifier when a concurrent claim stamps it first', async () => {
-      const patch = vi.fn().mockRejectedValueOnce(Object.assign(new Error('conflict'), { code: 412 }));
-      const get = vi.fn().mockResolvedValue({
-        ...billingResources.patient,
-        identifier: [clinicalPatientIdentifier('patient-123')],
-      });
-      await addClinicalPatientIdentifiers({
-        oystehr: {
-          fhir: {
-            patch,
-            get,
-          },
-        } as unknown as Oystehr,
-        patient: billingResources.patient,
-        clinicalId: 'patient-123',
-      });
-
-      expect(patch).toHaveBeenCalledTimes(1);
-      expect(get).toHaveBeenCalledWith({
-        resourceType: 'Patient',
-        id: 'billing-patient-123',
-      });
     });
   });
 });
